Build on msvc x86-64.

+++ components/xtl/tests/stl/accum1.cpp
@@ -10,7 +10,7 @@ int main ()
   
   vector<int> v (5);
   
-  for (size_t i=0;i<v.size();i++) v [i] = i+1;
+  for (size_t i=0;i<v.size();i++) v [i] = (int)i+1;
   
   printf ("sum = %d\n",accumulate (v.begin (),v.end (),0));
