+++ components/xtl/tests/stl/count0.cpp
@@ -7,7 +7,7 @@ int main ()
 {
   printf ("Results of count0_test:\n");
   
-  int numbers [10] = {1,2,4,1,2,4,1,2,4,1}, result = count (numbers,numbers+10,1);
+  int numbers [10] = {1,2,4,1,2,4,1,2,4,1}, result = (int)count (numbers,numbers+10,1);
 
   printf ("Found %d 1's.\n",result);
